+++ src/event/read.rs
@@ -53,12 +53,7 @@ impl InternalEventReader {
 
         let event_source = match self.source.as_mut() {
             Some(source) => source,
-            None => {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::Other,
-                    "Failed to initialize input reader",
-                ))
-            }
+            None => return Err(std::io::Error::other("Failed to initialize input reader")),
         };
 
         let poll_timeout = PollTimeout::new(timeout);
@@ -172,12 +167,16 @@ mod tests {
         };
 
         assert!(reader.poll(None, &InternalEventFilter).is_err());
-        assert!(reader
-            .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
-            .is_err());
-        assert!(reader
-            .poll(Some(Duration::from_secs(10)), &InternalEventFilter)
-            .is_err());
+        assert!(
+            reader
+                .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
+                .is_err()
+        );
+        assert!(
+            reader
+                .poll(Some(Duration::from_secs(10)), &InternalEventFilter)
+                .is_err()
+        );
     }
 
     #[test]
@@ -282,9 +281,11 @@ mod tests {
             skipped_events: Vec::with_capacity(32),
         };
 
-        assert!(!reader
-            .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
-            .unwrap());
+        assert!(
+            !reader
+                .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
+                .unwrap()
+        );
     }
 
     #[test]
@@ -298,9 +299,11 @@ mod tests {
         };
 
         assert!(reader.poll(None, &InternalEventFilter).unwrap());
-        assert!(reader
-            .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
-            .unwrap());
+        assert!(
+            reader
+                .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
+                .unwrap()
+        );
     }
 
     #[test]
@@ -350,9 +353,11 @@ mod tests {
         assert_eq!(reader.read(&InternalEventFilter).unwrap(), EVENT);
         assert_eq!(reader.read(&InternalEventFilter).unwrap(), EVENT);
         assert_eq!(reader.read(&InternalEventFilter).unwrap(), EVENT);
-        assert!(!reader
-            .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
-            .unwrap());
+        assert!(
+            !reader
+                .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
+                .unwrap()
+        );
     }
 
     #[test]
@@ -403,9 +408,11 @@ mod tests {
 
         assert_eq!(reader.read(&InternalEventFilter).unwrap(), EVENT);
         assert!(reader.read(&InternalEventFilter).is_err());
-        assert!(reader
-            .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
-            .unwrap());
+        assert!(
+            reader
+                .poll(Some(Duration::from_secs(0)), &InternalEventFilter)
+                .unwrap()
+        );
     }
 
     #[test]
@@ -435,7 +442,7 @@ mod tests {
         fn new(events: &[InternalEvent]) -> FakeSource {
             FakeSource {
                 events: events.to_vec().into(),
-                error: Some(io::Error::new(io::ErrorKind::Other, "")),
+                error: Some(io::Error::other("")),
             }
         }
 
@@ -450,11 +457,10 @@ mod tests {
     impl EventSource for FakeSource {
         fn try_read(&mut self, _timeout: Option<Duration>) -> io::Result<Option<InternalEvent>> {
             // Return error if set in case there's just one remaining event
-            if self.events.len() == 1 {
-                if let Some(error) = self.error.take() {
+            if self.events.len() == 1
+                && let Some(error) = self.error.take() {
                     return Err(error);
                 }
-            }
 
             // Return all events from the queue
             if let Some(event) = self.events.pop_front() {